refac set photo action and delete container
related #29
