--- src/client/actions/upload_photo.js
@@ -1,8 +1,0 @@
-import { SET_PHOTO } from '../action_types.js'
-
-export const setPhoto = photoData => {
-  return {
-    type: SET_PHOTO,
-    photoData
-  }
-}
